requirements.txt
azure-eventhub==5.11.0
requests==2.31.0
python-dotenv==1.0.1
